refactor(web): simplify about page components
